backend/requirements.txt
Django==4.2.30
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
openai==2.32.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
